+++ server/src-lib/Hasura/RQL/IR/Select.hs
@@ -24,10 +24,7 @@
 --     @UnpreparedValue b@ for their respective backend @b@, and most backends will then transform
 --     their AST, cutting all such remote branches, and therefore using @Const Void@ for @r@.
 module Hasura.RQL.IR.Select
-  ( ActionFieldG (..),
-    ActionFieldsG,
-    ActionFields,
-    AggregateField (..),
+  ( AggregateField (..),
     AggregateFields,
     AggregateOp (..),
     AnnAggregateSelect,
@@ -70,7 +67,6 @@ module Hasura.RQL.IR.Select
     ConnectionSplitKind (..),
     EdgeField (..),
     EdgeFields,
-    Fields,
     FunctionArgExp,
     FunctionArgsExpG (..),
     FunctionArgsExpTableRow,
@@ -146,7 +142,6 @@ import Hasura.RQL.Types.Instances ()
 import Hasura.RQL.Types.Relationships.Local
 import Hasura.RQL.Types.Relationships.Remote
 import Hasura.SQL.Backend
-import Language.GraphQL.Draft.Syntax qualified as G
 
 -- Root selection
 
@@ -401,10 +396,6 @@ type AnnotatedOrderByItemG b v = OrderByItemG b (AnnotatedOrderByElement b v)
 type AnnotatedOrderByItem b = AnnotatedOrderByItemG b (SQLExpression b)
 
 -- Fields
-
--- The field name here is the GraphQL alias, i.e, the name with which the field
--- should appear in the response
-type Fields a = [(FieldName, a)]
 
 -- | captures a remote relationship's selection and the necessary context
 data RemoteRelationshipSelect b r = RemoteRelationshipSelect
@@ -898,36 +889,6 @@ insertFunctionArg argName idx value (FunctionArgsExp positional named) =
         HM.insert (getFuncArgNameTxt argName) value named
   where
     insertAt i a = toList . Seq.insertAt i a . Seq.fromList
-
--- Actions
-
-data ActionFieldG (b :: BackendType) (r :: Type) v
-  = ACFScalar G.Name
-  | ACFObjectRelation (ObjectRelationSelectG b r v)
-  | ACFArrayRelation (ArraySelectG b r v)
-  | ACFExpression Text
-  | ACFNestedObject G.Name !(ActionFieldsG b r v)
-  deriving (Functor, Foldable, Traversable)
-
-deriving instance
-  ( Backend b,
-    Eq (BooleanOperators b v),
-    Eq v,
-    Eq r
-  ) =>
-  Eq (ActionFieldG b r v)
-
-deriving instance
-  ( Backend b,
-    Show (BooleanOperators b v),
-    Show v,
-    Show r
-  ) =>
-  Show (ActionFieldG b r v)
-
-type ActionFieldsG b r v = Fields (ActionFieldG b r v)
-
-type ActionFields b = ActionFieldsG b Void (SQLExpression b)
 
 -- | The "distinct" input field inside "count" aggregate field
 --