+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - server: fix parsing timestamp values in BigQuery backends (fix #8076)
 - server: add support for customising the GraphQL schema descriptions of table root fields
 - server: add a `request_headers` field to the `test_webhook_transform` API.
+- server: add support for relationships on nested action fields
 - console: include cron trigger with include in metadata as false on cron trigger manage page
 - console: show an error notification if Hasura CLI migrations fail
 - console: fixed an issue where cron triggers were not removed from the list after deletion